requirements.txt
numpy>=1.24
pillow>=10.0
pyyaml>=6.0
scikit-learn>=1.3

# PyTorch: use ONE of the following.
# Intel GPU (XPU) — recommended for this project:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/xpu
# CPU / CUDA — see https://pytorch.org/get-started/locally/
torch>=2.4
torchvision>=0.19
